requirements/local_django.txt
-r ./base_django.txt

# Testing
pytest==8.1.1
pytest-sugar==1.0.0
pytest-django==4.8.0

# Debugging
ipdb==0.13.13